add some verifications before and after build

+++ scripts/buildmsi.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python
-import sys,os,re,uuid,subprocess
+import sys,os,re,uuid,subprocess,shutil
 
 #
 # create MSI file for installation of OpenSCAD program on Microsoft Windows
 # computer operating system. from inside of a linux cross-build environment.
+# using Gnome's Wixl and Wixl-Heat which are based on the 'WiX' project.
 #
 # why:
 #
@@ -57,24 +58,77 @@ import sys,os,re,uuid,subprocess
 # https://wiki.gnome.org/msitools/HowTo/CreateLibraryWxi
 # wixl test suite (msitools 'make check-local')
 
-def main(openscad_crossbuild_dir, openscad_srcdir, openscad_version, arch ):
+def verify_deps():
+	score = 0
+	for exe in ['msiinfo','wixl-heat','wixl']:
+		check = shutil.which(exe)
+		if check!=None: score += 1
+		print(exe,'\t', check)
+	return score>0
+
+def verify_msi(msi_filename):
+	if os.path.exists(msi_filename):
+		print('created',msi_filename,'size',os.path.getsize(msi_filename))
+	else:
+		print('sorry something went awry. no',msi_filename,'created')
+		return False
+
+	cmds = [['msiinfo','export',msi_filename,'File'],
+	        ['msiinfo','export',msi_filename,'Directory']]
+	lines = []
+	for cmd in cmds:
+		print('running',' '.join(cmd))
+		p=subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+		lines += p.stdout.readlines() + p.stderr.readlines()
+	sanity = [0,0]
+	for line in lines:
+		l = line.decode('utf-8').strip().replace('\t',' ')[:79]
+		if 'openscad_executable' in l:
+			sanity[0]=1
+			print(l)
+		if 'INSTALLDIR' in l:
+			sanity[1]=1
+			print(l)
+	if sanity[0]+sanity[1]<2:
+		print('sorry something went awry.',msi_filename,'appears')
+		print('to be missing openscad.exe and/or a proper INSTALLDIR')
+		return False
+	return True
+
+def verify_path(dir,file):
+	if not os.path.exists(os.path.join(dir,file)):
+		print('cannot find',os.path.join(dir,file))
+		return False
+	return True
+
+def main(openscad_crossbuild_dir, openscad_src_dir, openscad_version, arch ):
 	print('build MSI for openscad')
 	print('openscad_crossbuild_dir',openscad_crossbuild_dir)
-	print('openscad_srcdir',openscad_srcdir)
+	print('openscad_src_dir',openscad_src_dir)
 	print('openscad_version',openscad_version)
 	print('architecture', arch)
+
+	wixarchcodes = {'x86-32':'intel','x86-64':'x64'}
+
+	print('please cross-build before running this script')
+	if not verify_path(openscad_src_dir,'README.md'):
+		return
+	if not verify_path(openscad_crossbuild_dir,'openscad.exe'):
+		print('please cross-build openscad.exe before running this script')
+		return
+	if not arch in wixarchcodes.keys():
+		print('cannot find arch',arch,'in ',wixarchcodes)
+		return
 
 	mainwxs_filename = 'openscad.wxs'
 	filelistwxs_filename = 'filelist.wxs'
 	msi_filename = 'openscad.msi'
 
 	if 'clean' in ''.join(sys.argv[1:]):
-		for f in 'filelist.txt',filelistwxs_filename,msi_filename:
+		for f in filelistwxs_filename,msi_filename:
 			if os.path.exists(f): os.remove(f)
 			print('clean',f)
 		return
-
-	wixarchcodes = {'x86-32':'intel','x86-64':'x64'}
 
 	filelist=''
 	for root,folder,files in os.walk(openscad_crossbuild_dir):
@@ -119,7 +173,7 @@ def main(openscad_crossbuild_dir, openscad_srcdir, openscad_version, arch ):
 	cmd=['wixl','--verbose']
 	cmd+=['--arch',wixarchcodes[arch]]
 	cmd+=['--define','OPENSCADCROSSBUILDDIR='+openscad_crossbuild_dir]
-	cmd+=['--define','OPENSCADSRCDIR='+openscad_srcdir]
+	cmd+=['--define','OPENSCADsrc_dir='+openscad_src_dir]
 	cmd+=['--define','OPENSCADVERSION='+openscad_version]
 	cmd+=['--output',msi_filename]
 	cmd+=[mainwxs_filename,filelistwxs_filename]
@@ -128,32 +182,12 @@ def main(openscad_crossbuild_dir, openscad_srcdir, openscad_version, arch ):
 	print(p.stdout.read().decode('utf-8'))
 	print(p.stderr.read().decode('utf-8'))
 
-	if os.path.exists(msi_filename):
-		print('created',msi_filename,'size',os.path.getsize(msi_filename))
-	else:
-		print('sorry something went awry. no',msi_filename,'created')
-		return
+	verify_msi()
 
-	cmds = [['msiinfo','export',msi_filename,'File'],
-	        ['msiinfo','export',msi_filename,'Directory']]
-	lines = []
-	for cmd in cmds:
-		print('running',' '.join(cmd))
-		p=subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-		lines += p.stdout.readlines() + p.stderr.readlines()
-	sanity = [0,0]
-	for line in lines:
-		l = line.decode('utf-8').strip().replace('\t',' ')[:79]
-		if 'openscad_executable' in l:
-			sanity[0]=1
-			print(l)
-		if 'INSTALLDIR' in l:
-			sanity[1]=1
-			print(l)
-	if sanity[0]+sanity[1]<2:
-		print('sorry something went awry.',msi_filename,'appears')
-		print('to be missing openscad.exe and/or a proper INSTALLDIR')
-		return
-
-main('./openscad32','/home/don/src/openscad','2018.08.12','x86-32')
-#main('./openscad64','/home/don/src/openscad','2018.08.12','x86-64')
+if verify_deps():
+	main('./openscad32','/home/don/src/openscad','2018.08.12','x86-32')
+	#main('./openscad64','/home/don/src/openscad','2018.08.12','x86-64')
+else:
+	print("please install the MSI tools for your platform")
+	print("i need msiinfo, wixl, and wixl-heat")
+	print("please see https://wiki.gnome.org/msitools")
